+++ server/src-lib/Hasura/RQL/Types/Metadata.hs
@@ -11,6 +11,7 @@ import qualified Data.HashSet                        as HS
 import qualified Data.HashSet.InsOrd                 as HSIns
 import qualified Data.List.Extended                  as L
 import qualified Data.Text                           as T
+import qualified Data.Text.Extended                  as T
 import qualified Language.GraphQL.Draft.Syntax       as G
 
 import           Control.Lens                        hiding (set, (.=))
@@ -34,6 +35,7 @@ import           Hasura.RQL.Types.EventTrigger
 import           Hasura.RQL.Types.Function
 import           Hasura.RQL.Types.InheritedRoles
 import           Hasura.RQL.Types.Metadata.Backend
+import           Hasura.RQL.Types.Metadata.Instances ()
 import           Hasura.RQL.Types.Permission
 import           Hasura.RQL.Types.QueryCollection
 import           Hasura.RQL.Types.Relationship
@@ -42,7 +44,9 @@ import           Hasura.RQL.Types.RemoteSchema
 import           Hasura.RQL.Types.ScheduledTrigger
 import           Hasura.RQL.Types.Table
 import           Hasura.SQL.Backend
+import           Hasura.SQL.Tag
 import           Hasura.Session
+
 
 -- | Raise exception if parsed list has multiple declarations
 parseListAsMap
@@ -465,7 +469,7 @@ metadataToOrdJSON ( Metadata
     sourceMetaToOrdJSON exists =
       AB.dispatchAnyBackend @BackendMetadata exists $ \(SourceMetadata {..} :: SourceMetadata b) ->
         let sourceNamePair = ("name", AO.toOrdered _smName)
-            sourceKind     = backendName $ backendTag @b
+            sourceKind     = T.toTxt $ reify $ backendTag @b
             sourceKindPair = ("kind", AO.String sourceKind)
             tablesPair     = ("tables", AO.array $ map tableMetaToOrdJSON $ sortOn _tmTable $ OM.elems _smTables)
             functionsPair  = listToMaybeOrdPairSort "functions" functionMetadataToOrdJSON _fmFunction _smFunctions